pyproject.toml
[project]
name = "reflectiondirection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.4.0",
    "datasets>=3.3.2",
    "jupyter>=1.1.1",
    "nupy>=0.1.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "transformers-stream-generator>=0.0.5",
    "vllm>=0.7.3",
]
